requirements.txt
osu_py==4.1.0
python-dotenv==1.2.2
Requests==2.33.1
